Use recommended module structure
