+++ presentation/lib/src/home_page.dart
@@ -1,0 +1,119 @@
+import 'package:domain/domain.dart';
+import 'package:flutter/material.dart';
+import 'package:presentation/src/keys.dart';
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({
+    super.key,
+    required this.title,
+    required CounterRepository counterRepository,
+  }) : _counterRepository = counterRepository;
+
+  final String title;
+  final CounterRepository _counterRepository;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  Counter? _counter;
+
+  @override
+  void initState() {
+    super.initState();
+    _initialiseCounter();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text(
+              "The counter's value is:",
+            ),
+            Text(
+              _counter?.value == null ? 'Not yet loaded' : '${_counter!.value}',
+              style: Theme.of(context).textTheme.headlineMedium,
+            ),
+            ButtonBar(
+              alignment: MainAxisAlignment.center,
+              children: [
+                TextButton(
+                  key: Keys.decrement.value,
+                  onPressed: _decrementCounter,
+                  style: ButtonStyle(
+                    backgroundColor: MaterialStatePropertyAll(
+                      Theme.of(context).colorScheme.primaryContainer,
+                    ),
+                    overlayColor: MaterialStatePropertyAll(
+                      Theme.of(context).colorScheme.inversePrimary,
+                    ),
+                  ),
+                  child: Text(
+                    '-',
+                    style: Theme.of(context).textTheme.headlineSmall,
+                  ),
+                ),
+                TextButton(
+                  key: Keys.increment.value,
+                  onPressed: _incrementCounter,
+                  style: ButtonStyle(
+                    backgroundColor: MaterialStatePropertyAll(
+                      Theme.of(context).colorScheme.primaryContainer,
+                    ),
+                    overlayColor: MaterialStatePropertyAll(
+                      Theme.of(context).colorScheme.inversePrimary,
+                    ),
+                  ),
+                  child: Text(
+                    '+',
+                    style: Theme.of(context).textTheme.headlineSmall,
+                  ),
+                )
+              ],
+            )
+          ],
+        ),
+      ),
+    );
+  }
+
+  Future<void> _initialiseCounter() async {
+    final retrievedCounter = await widget._counterRepository.get();
+
+    setState(() {
+      _counter = retrievedCounter ?? Counter();
+    });
+  }
+
+  Future<void> _incrementCounter() async {
+    if (_counter == null) {
+      return;
+    }
+
+    _counter!.increment();
+    await widget._counterRepository.save(_counter!);
+
+    setState(() {});
+  }
+
+  Future<void> _decrementCounter() async {
+    if (_counter == null) {
+      return;
+    }
+
+    if (_counter!.tryDecrement()) {
+      await widget._counterRepository.save(_counter!);
+
+      setState(() {});
+    }
+  }
+}
