+++ presentation/lib/src/keys.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/foundation.dart';
+
+enum Keys {
+  increment(ValueKey("increment")),
+  decrement(ValueKey("decrement"));
+
+  final ValueKey value;
+
+  const Keys(this.value);
+}
